requirements.txt
Django>=5.0
djangorestframework>=3.14
pandas
openpyxl
reportlab
psycopg2-binary
gunicorn
rest_franmework